pyproject.toml
[tool.poetry]
name = "ulauncher-toggl-extension"
version = "2.1.3"
description = "Toggl time tracker extension for Ulauncher."
authors = ["David Kasakaitis <davidkasakaitis@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
toggl-api-wrapper = "^0.3.0"
pycairo = "^1.26.1"
pygobject = "^3.48.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pygobject-stubs = "^2.10.0"
pytest-cov = "^5.0.0"
tox-gh-actions = "^3.2.0"
ruff = "^0.5.7"
mypy = "^1.11.1"
faker = "^27.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
exclude = []

[tool.ruff.lint]
preview = true
select = [
    "F",
    "E",
    "W",
    "C90",
    "I",
    "N",
    "UP",
    "YTT",
    "ASYNC",
    "S",
    "BLE",
    "FBT",
    "B",
    "A",
    "COM",
    "C4",
    "DTZ",
    "T10",
    "DJ",
    "EM",
    "EXE",
    "FA",
    'ISC',
    "ICN",
    "G",
    'INP',
    'PIE',
    "T20",
    'PYI',
    'PT',
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SLOT",
    "SIM",
    "TID",
    "TCH",
    "INT",
    "ARG",
    "PTH",
    "ERA",
    "PD",
    "PGH",
    "PL",
    "TRY",
    "FLY",
    "NPY",
    "AIR",
    "PERF",
    "FURB",
    "LOG",
    "RUF",
]
ignore = ["S101", "RUF012", "SIM102", "UP007", "PLW2901"]

[tool.mypy]
ignore_missing_imports = true
disable_error_code = ["attr-defined", "assignment"]
exclude = ["tests/", "main.py", "__init__.py"]

[tool.pytest.ini_options]
markers = [
    "unit: standard tests that don't access external resources.",
    "integration: tests that access external apis.",
]
testpaths = ["tests"]

[tool.tox]
legacy_tox_ini = """
[tox]
isolated_build = True
envlist = ruff, mypy, py312, py311, py310

[testenv]
skip_install = true
passenv = *
setenv = 
    GH_ACTION = "True"
    TOGGL_WORKSPACE_ID = {env:TOGGL_WORKSPACE_ID}
    TOGGL_API_TOKEN = {env:TOGGL_API_TOKEN}
    PYTHONPATH = {toxinidir}

allowlist_externals = poetry
commands_pre = 
    poetry install
commands = 
    poetry run pytest {posargs}

[gh-actions]
python = 
    3.10: py310
    3.11: py311
    3.12: ruff, mypy, py312

[testenv:mypy]
allowlist_externals = mypy, poetry
basepython = python3.12
commands = mypy ulauncher_toggl_extension/

[testenv:ruff]
allowlist_externals = ruff, poetry
basepython = python3.12
commands = ruff check ulauncher_toggl_extension/
"""
